refactor database into multiple files
