+++ src/tools/web/clusterd/src/db/clusters.js
@@ -1,70 +1,6 @@
-import kdb from '../../kdb'
-import { v4 as generateId } from 'node-uuid'
+import kdb from '../../../kdb'
+import { path, generateId, findById, updateById, deleteById } from './utils'
 
-const ROOT_PATH = 'user/sw/elektra/web/#0/current'
-
-const path = (path) =>
-  ROOT_PATH + '/' + path
-
-const findById = (id) =>
-  elements => elements.find(
-    element => element.id === id
-  )
-
-const updateById = (id, update) =>
-  elements => elements.reduce(
-    (acc, element) => {
-      if (element.id === id) {
-        acc.push(update(element))
-      } else {
-        acc.push(element)
-      }
-      return acc
-    }, []
-  )
-
-const deleteById = (id) =>
-  elements => elements.filter(
-    element => element.id !== id
-  )
-
-// instances
-export const getInstances = () =>
-  kdb.export(path('instances'))
-    .then(res => Array.isArray(res) ? res : [])
-
-const persistInstances = (instances) =>
-  kdb.import(path('instances'), instances)
-
-export const createInstance = ({ name, host }) => {
-  const generatedId = generateId()
-  return getInstances()
-    .then(instances => instances.concat({ id: generatedId, name, host }))
-    .then(persistInstances)
-    .then(findById(generatedId))
-}
-
-export const getInstance = (id) =>
-  getInstances()
-    .then(findById(id))
-
-export const updateInstance = (id, data) =>
-  getInstances()
-    .then(updateById(id, (instance) => {
-      return { ...instance, ...data }
-    }))
-    .then(persistInstances)
-    .then(findById(id))
-
-export const deleteInstance = (id) =>
-  getInstances()
-    .then(deleteById(id))
-    .then(persistInstances)
-    .then(instances => {
-      return { id } // return id of deleted instance
-    })
-
-// clusters
 export const getClusters = () =>
   kdb.export(path('clusters'))
     .then(res => Array.isArray(res) ? res : [])
