+++ src/tools/web/clusterd/src/db/index.js
@@ -1,0 +1,8 @@
+export {
+  getInstances, createInstance, getInstance, updateInstance, deleteInstance
+} from './instances'
+
+export {
+  getClusters, createCluster, getCluster, updateCluster, deleteCluster,
+  getClusterInstances, addInstanceToCluster, removeInstanceFromCluster
+} from './clusters'
